Hide and fix some warnings.
Enable MP build.

We can hide some c4251 warnings till we translate sd2 to static.

This is for VS, we'll be happy if someone can do the same for nix

+++ src/game/SpellEffects.cpp
@@ -1579,7 +1579,7 @@ void Spell::EffectPowerDrain(SpellEffectIndex eff_idx)
     damage = unitTarget->SpellDamageBonusTaken(m_caster, m_spellInfo, uint32(damage), SPELL_DIRECT_DAMAGE);
 
     int32 new_damage;
-    if (curPower < damage)
+    if (curPower < uint32(damage)) // damage should not be under zero at this point (checked above)
         new_damage = curPower;
     else
         new_damage = damage;
